+++ src/liquid/components/ld-button/test/ld-button.e2e.ts
@@ -1,8 +1,9 @@
-import { getPageWithContent } from '../../../utils/e2e-tests'
+import {
+  analyzeAccessibility,
+  getPageWithContent,
+} from '../../../utils/e2e-tests'
 import { LdIcon } from '../../ld-icon/ld-icon'
 import { LdButton } from '../ld-button'
-
-jest.useRealTimers()
 
 const modes = ['', 'highlight', 'secondary', 'ghost', 'danger']
 const attributes = ['', 'brand-color']
@@ -44,6 +45,9 @@ describe('ld-button', () => {
           )
           const results = await page.compareScreenshot()
           expect(results).toMatchScreenshot()
+
+          const accessibilityReport = await analyzeAccessibility(page)
+          expect(accessibilityReport).toHaveNoAccessibilityIssues()
         })
         it('hover', async () => {
           const page = await getPageWithContent(
@@ -69,6 +73,7 @@ describe('ld-button', () => {
             }
           )
           await page.keyboard.press('Tab')
+          await page.waitForChanges()
           const results = await page.compareScreenshot()
           expect(results).toMatchScreenshot()
         })
@@ -83,6 +88,7 @@ describe('ld-button', () => {
           )
           await page.keyboard.press('Tab')
           await page.keyboard.down('Space')
+          await page.waitForChanges()
           const results = await page.compareScreenshot()
           expect(results).toMatchScreenshot()
         })
@@ -99,6 +105,9 @@ describe('ld-button', () => {
             )
             const results = await page.compareScreenshot()
             expect(results).toMatchScreenshot()
+
+            const accessibilityReport = await analyzeAccessibility(page)
+            expect(accessibilityReport).toHaveNoAccessibilityIssues()
           })
           it('hover', async () => {
             const page = await getPageWithContent(
@@ -124,6 +133,7 @@ describe('ld-button', () => {
               }
             )
             await page.keyboard.press('Tab')
+            await page.waitForChanges()
             const results = await page.compareScreenshot()
             expect(results).toMatchScreenshot()
           })
@@ -138,6 +148,7 @@ describe('ld-button', () => {
             )
             await page.keyboard.press('Tab')
             await page.keyboard.down('Space')
+            await page.waitForChanges()
             const results = await page.compareScreenshot()
             expect(results).toMatchScreenshot()
           })
@@ -180,6 +191,7 @@ describe('ld-button', () => {
               }
             )
             await page.keyboard.press('Tab')
+            await page.waitForChanges()
             const results = await page.compareScreenshot()
             expect(results).toMatchScreenshot()
           })
@@ -194,6 +206,7 @@ describe('ld-button', () => {
             )
             await page.keyboard.press('Tab')
             await page.keyboard.down('Space')
+            await page.waitForChanges()
             const results = await page.compareScreenshot()
             expect(results).toMatchScreenshot()
           })
@@ -239,6 +252,7 @@ describe('ld-button', () => {
               }
             )
             await page.keyboard.press('Tab')
+            await page.waitForChanges()
             const results = await page.compareScreenshot()
             expect(results).toMatchScreenshot()
           })
@@ -254,6 +268,7 @@ describe('ld-button', () => {
             )
             await page.keyboard.press('Tab')
             await page.keyboard.down('Space')
+            await page.waitForChanges()
             const results = await page.compareScreenshot()
             expect(results).toMatchScreenshot()
           })
@@ -297,6 +312,7 @@ describe('ld-button', () => {
               }
             )
             await page.keyboard.press('Tab')
+            await page.waitForChanges()
             const results = await page.compareScreenshot()
             expect(results).toMatchScreenshot()
           })
@@ -310,6 +326,7 @@ describe('ld-button', () => {
             )
             await page.keyboard.press('Tab')
             await page.keyboard.down('Space')
+            await page.waitForChanges()
             const results = await page.compareScreenshot()
             expect(results).toMatchScreenshot()
           })
@@ -348,6 +365,7 @@ describe('ld-button', () => {
                 }
               )
               await page.keyboard.press('Tab')
+              await page.waitForChanges()
               const results = await page.compareScreenshot()
               expect(results).toMatchScreenshot()
             })
@@ -361,6 +379,7 @@ describe('ld-button', () => {
               )
               await page.keyboard.press('Tab')
               await page.keyboard.down('Space')
+              await page.waitForChanges()
               const results = await page.compareScreenshot()
               expect(results).toMatchScreenshot()
             })
@@ -400,6 +419,7 @@ describe('ld-button', () => {
                 }
               )
               await page.keyboard.press('Tab')
+              await page.waitForChanges()
               const results = await page.compareScreenshot()
               expect(results).toMatchScreenshot()
             })
@@ -413,6 +433,7 @@ describe('ld-button', () => {
               )
               await page.keyboard.press('Tab')
               await page.keyboard.down('Space')
+              await page.waitForChanges()
               const results = await page.compareScreenshot()
               expect(results).toMatchScreenshot()
             })
@@ -464,6 +485,7 @@ describe('ld-button', () => {
                 }
               )
               await page.keyboard.press('Tab')
+              await page.waitForChanges()
               const results = await page.compareScreenshot()
               expect(results).toMatchScreenshot()
             })
@@ -481,6 +503,7 @@ describe('ld-button', () => {
               )
               await page.keyboard.press('Tab')
               await page.keyboard.down('Space')
+              await page.waitForChanges()
               const results = await page.compareScreenshot()
               expect(results).toMatchScreenshot()
             })
@@ -497,6 +520,9 @@ describe('ld-button', () => {
       )
       const results = await page.compareScreenshot()
       expect(results).toMatchScreenshot()
+
+      const accessibilityReport = await analyzeAccessibility(page)
+      expect(accessibilityReport).toHaveNoAccessibilityIssues()
     })
     it('lg', async () => {
       const page = await getPageWithContent(
@@ -517,10 +543,13 @@ describe('ld-button', () => {
     })
     it('icon button sm', async () => {
       const page = await getPageWithContent(
-        '<ld-button size="sm"><ld-sr-only>Text</ld-sr-only><ld-icon name="placeholder"></ld-icon></ld-button>'
+        '<ld-button size="sm"><ld-icon name="placeholder"></ld-icon><ld-sr-only>Text</ld-sr-only></ld-button>'
       )
       const results = await page.compareScreenshot()
       expect(results).toMatchScreenshot()
+
+      const accessibilityReport = await analyzeAccessibility(page)
+      expect(accessibilityReport).toHaveNoAccessibilityIssues()
     })
     it('icon button lg', async () => {
       const page = await getPageWithContent(
@@ -537,6 +566,7 @@ describe('ld-button', () => {
         )
         await page.keyboard.press('Tab')
         await page.keyboard.down('Space')
+        await page.waitForChanges()
         const results = await page.compareScreenshot()
         expect(results).toMatchScreenshot()
       })
@@ -546,6 +576,7 @@ describe('ld-button', () => {
         )
         await page.keyboard.press('Tab')
         await page.keyboard.down('Space')
+        await page.waitForChanges()
         const results = await page.compareScreenshot()
         expect(results).toMatchScreenshot()
       })
@@ -555,6 +586,7 @@ describe('ld-button', () => {
         )
         await page.keyboard.press('Tab')
         await page.keyboard.down('Space')
+        await page.waitForChanges()
         const results = await page.compareScreenshot()
         expect(results).toMatchScreenshot()
       })
@@ -599,6 +631,9 @@ describe('ld-button', () => {
       )
       const results = await page.compareScreenshot()
       expect(results).toMatchScreenshot()
+
+      const accessibilityReport = await analyzeAccessibility(page)
+      expect(accessibilityReport).toHaveNoAccessibilityIssues()
     })
     it('multiline text left', async () => {
       const page = await getPageWithContent(
@@ -623,6 +658,9 @@ describe('ld-button', () => {
       )
       const results = await page.compareScreenshot()
       expect(results).toMatchScreenshot()
+
+      const accessibilityReport = await analyzeAccessibility(page)
+      expect(accessibilityReport).toHaveNoAccessibilityIssues()
     })
   })