+++ src/liquid/components/ld-input/ld-input.tsx
@@ -73,6 +73,9 @@ export class LdInput implements InnerFocusable, ClonesAttributes {
   /** Set this property to `true` in order to mark the field visually as invalid. */
   @Prop() invalid?: boolean
 
+  /** Tab index of the input. */
+  @Prop() ldTabindex: number | undefined
+
   /** Value of the id attribute of the `<datalist>` of autocomplete options. */
   @Prop() list?: string
 
@@ -111,6 +114,9 @@ export class LdInput implements InnerFocusable, ClonesAttributes {
 
   /** A value is required for the form to be submittable. */
   @Prop() required?: boolean
+
+  /** Whether the multiline input is resizable, and if so, in which directions. */
+  @Prop() resize?: 'none' | 'both' | 'horizontal' | 'vertical' = 'both'
 
   /** The number of rows. */
   @Prop() rows?: number
@@ -332,9 +338,11 @@ export class LdInput implements InnerFocusable, ClonesAttributes {
   render() {
     const cl = getClassNames([
       'ld-input',
+      this.disabled && `ld-input--disabled`,
       this.size && `ld-input--${this.size}`,
       this.tone && `ld-input--${this.tone}`,
       this.invalid && 'ld-input--invalid',
+      this.multiline && this.resize && `ld-input--resize-${this.resize}`,
     ])
 
     if (this.multiline) {
@@ -347,6 +355,7 @@ export class LdInput implements InnerFocusable, ClonesAttributes {
             onInput={this.handleInput}
             part="input focusable"
             ref={(el) => (this.input = el)}
+            tabIndex={this.ldTabindex}
             value={this.value}
           />
           {type === 'file' && (
@@ -369,6 +378,7 @@ export class LdInput implements InnerFocusable, ClonesAttributes {
           onKeyDown={this.handleKeyDown}
           part="input focusable"
           ref={(el) => (this.input = el)}
+          tabIndex={this.ldTabindex}
           value={this.value}
         />
         {this.type === 'file' && (