+++ src/liquid/components/ld-select/ld-select.tsx
@@ -57,6 +57,9 @@ export class LdSelect implements InnerFocusable {
   /** Set this property to `true` in order to mark the select visually as invalid. */
   @Prop() invalid: boolean
 
+  /** Tab index of the trigger button. */
+  @Prop() ldTabindex = 0
+
   /** Constrains the height of the trigger button by replacing overflowing selection with a "+X more" indicator. */
   @Prop({ mutable: true }) maxRows?: number
 
@@ -895,6 +898,7 @@ export class LdSelect implements InnerFocusable {
 
     const cl = [
       'ld-select',
+      this.disabled && 'ld-select--disabled',
       this.size && `ld-select--${this.size}`,
       this.invalid && 'ld-select--invalid',
       this.expanded && 'ld-select--expanded',
@@ -949,7 +953,11 @@ export class LdSelect implements InnerFocusable {
               class={getClassNames(triggerCl)}
               role="button"
               part="btn-trigger focusable"
-              tabindex={this.disabled && !this.ariaDisabled ? undefined : '0'}
+              tabindex={
+                this.disabled && !this.ariaDisabled
+                  ? undefined
+                  : this.ldTabindex
+              }
               aria-disabled={
                 this.disabled || this.ariaDisabled ? 'true' : 'false'
               }