+++ src/liquid/components/ld-tabs/ld-tab/ld-tab.tsx
@@ -29,14 +29,17 @@ export class LdTab implements InnerFocusable {
   private btnRef: HTMLButtonElement
 
   /**
-   * If present, this boolean attribute indicates that the tab is selected.
-   */
-  @Prop({ mutable: true, reflect: true }) selected?: boolean
-
-  /**
    * Disables the tab.
    */
   @Prop() disabled?: boolean
+
+  /** Tab index of the tab. */
+  @Prop() ldTabindex: number | undefined
+
+  /**
+   * If present, this boolean attribute indicates that the tab is selected.
+   */
+  @Prop({ mutable: true, reflect: true }) selected?: boolean
 
   /**
    * Focuses the tab
@@ -84,7 +87,7 @@ export class LdTab implements InnerFocusable {
         part="button focusable"
         ref={(el) => (this.btnRef = el as HTMLButtonElement)}
         role="tab"
-        tabindex={this.selected ? undefined : '-1'}
+        tabindex={this.selected ? this.ldTabindex : '-1'}
       >
         <span class="ld-tab__spacer" part="spacer spacer-left" />
         <span class="ld-tab__content" part="content">