+++ stencil.config.ts
@@ -18,6 +18,7 @@ export const config: Config = {
     },
     {
       type: 'dist-custom-elements',
+      autoDefineCustomElements: true,
     },
     {
       type: 'docs-readme',